Updating BrokerProperties and Message properties
json field is named LockedUntilUtc
TimeToLive is floating-point

+++ microsoft-azure-api/src/main/java/com/microsoft/azure/services/serviceBus/Message.java
@@ -89,11 +89,11 @@ public class Message {
 		return properties.getSequenceNumber();
 	}
 
-	public Long getTimeToLive() {
+	public Double getTimeToLive() {
 		return properties.getTimeToLive();
 	}
 
-	public Message setTimeToLive(Long timeToLive) {
+	public Message setTimeToLive(Double timeToLive) {
 		properties.setTimeToLive(timeToLive);
 		return this;
 	}
